# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the professional backend API for the Echo messaging platform.",
        "title": "Echo Messenger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Moeed-ul-Hassan",
            "url": "http://www.moeed.me",
            "email": "moeed@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8001",
    "basePath": "/",
    "paths": {
...
        "/api/stories/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "stories"
                ],
                "summary": "List active stories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/story.Story"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stories"
                ],
                "summary": "Create story",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/story.Story"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/stories/{storyID}/view": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "stories"
                ],
                "summary": "Mark story viewed",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Story ID",
                        "name": "storyID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/stories/{storyID}/viewers": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "stories"
                ],
                "summary": "Get story viewers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Story ID",
                        "name": "storyID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/story.Viewer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/user/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns the full profile of the authenticated user based on their PASETO token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get 'Me' Profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    }
                }
            }
        },
        "/api/user/search": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Finds a public profile of another user by their @handle.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Search User by Username",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User handle to search for",
                        "name": "username",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "story.Story": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "description": "Set to CreatedAt + 24h",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "media_type": {
                    "description": "image, video",
                    "type": "string"
                },
                "media_url": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "viewers": {
                    "description": "Users who swiped up/viewed",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/story.Viewer"
                    }
                }
            }
        },
        "story.Viewer": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string"
                },
                "viewed_at": {
                    "type": "string"
                }
            }
        },
        "user.User": {
            "type": "object",
            "properties": {
                "bio": {
                    "description": "Short bio or status",
                    "type": "string"
                },
                "clerk_id": {
                    "description": "Link to Clerk user",
                    "type": "string"
                },
                "country": {
                    "description": "Country detected via IP",
                    "type": "string"
                },
                "created_at": {
                    "description": "Creation timestamp",
                    "type": "string"
                },
                "email": {
                    "description": "Verified email address",
                    "type": "string"
                },
                "id": {
                    "description": "Unique UUID string",
                    "type": "string"
                },
                "is_admin": {
                    "description": "Admin flag",
                    "type": "boolean"
                },
                "is_verified": {
                    "description": "Has user verified email?",
                    "type": "boolean"
                },
                "last_seen": {
                    "description": "Last interaction timestamp",
                    "type": "string"
                },
                "name": {
                    "description": "Full display name",
                    "type": "string"
                },
                "status": {
                    "description": "Current status message",
                    "type": "string"
                },
                "username": {
                    "description": "Human-readable handle",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}